refactor(lambda): converts to aws-lambda-nodejs (#113)
