+++ .projenrc.js
@@ -23,8 +23,24 @@ const project = new AwsCdkConstructLibrary({
     '@aws-cdk/custom-resources',
     '@aws-cdk/aws-iam',
     '@aws-cdk/aws-lambda',
-    '@aws-cdk/aws-lambda-python'
+    '@aws-cdk/aws-lambda-nodejs'
   ],
+
+  devDeps: [
+    '@types/aws-lambda',
+    'esbuild@^0'
+  ],
+
+  deps: [
+    'aws-lambda',
+    'aws-sdk',
+  ],
+
+  bundledDeps: [
+    'aws-lambda',
+    'aws-sdk',
+  ],
+
   codeCov: true,
 
   python: {
