--- resources/lambda/crhelper-2.0.6.dist-info/INSTALLER
@@ -1,1 +1,0 @@
-pip
