--- resources/lambda/crhelper-2.0.6.dist-info/WHEEL
@@ -1,5 +1,0 @@
-Wheel-Version: 1.0
-Generator: bdist_wheel (0.34.2)
-Root-Is-Purelib: true
-Tag: py3-none-any
-
