--- resources/lambda/crhelper-2.0.6.dist-info/top_level.txt
@@ -1,2 +1,0 @@
-crhelper
-tests
