--- resources/lambda/crhelper/__init__.py
@@ -1,1 +1,0 @@
-from crhelper.resource_helper import CfnResource, SUCCESS, FAILED
